refactor: 기존 asis가 없어도 작동할 수 있도록 변경
