+++ app/src/main/java/com/techcourse/controller/IndexController.java
@@ -1,0 +1,16 @@
+package com.techcourse.controller;
+
+import context.org.springframework.stereotype.Controller;
+import web.org.springframework.web.bind.annotation.RequestMapping;
+import web.org.springframework.web.bind.annotation.RequestMethod;
+import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.view.JspView;
+
+@Controller
+public class IndexController {
+
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public ModelAndView getIndexView() {
+        return new ModelAndView(new JspView("/index.jsp"));
+    }
+}
